refs #81: refactor package
